requirements.txt
#apache-airflow-providers-google==10.3.0
astronomer-cosmos[dbt-bigquery]>=1.0.3
protobuf>=3.6.0
apache-airflow-providers-dbt-cloud==2.2.0
python-dotenv
pandas